# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/accounts": {
      "post": {
        "operationId": "AccountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccount"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "operationId": "AccountsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/me": {
      "get": {
        "operationId": "AccountsController_findMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "operationId": "AccountsController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBusiness"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "operationId": "AccountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      },
      "delete": {
        "operationId": "AccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    }
  },
  "info": {
    "title": "Movu API",
    "description": "Backend for all movu services",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "clerkSession": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Use a Clerk session token (Authorization: Bearer <token>)."
      }
    },
    "schemas": {
      "CreateAccount": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "BUSINESS",
              "PERSONAL"
            ]
          }
        },
        "required": [
          "userId",
          "kind"
        ],
        "id": "CreateAccount"
      },
      "RegisterBusiness": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1
          },
          "country": {
            "type": "string",
            "minLength": 1
          },
          "businessName": {
            "type": "string",
            "minLength": 1
          },
          "businessRegistrationNumber": {
            "type": "string",
            "minLength": 1
          },
          "taxId": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "postalCode": {
            "type": "string",
            "minLength": 1
          },
          "city": {
            "type": "string",
            "minLength": 1
          },
          "tradeRole": {
            "type": "string",
            "enum": [
              "IMPORTER",
              "EXPORTER",
              "FREIGHT_FORWARDER",
              "CUSTOMS_BROKER",
              "MANUFACTURER",
              "DISTRIBUTOR",
              "TRADER"
            ]
          }
        },
        "required": [
          "fullName",
          "email",
          "phoneNumber",
          "country",
          "businessName",
          "businessRegistrationNumber",
          "taxId",
          "address",
          "postalCode",
          "city",
          "tradeRole"
        ],
        "id": "RegisterBusiness"
      },
      "UpdateAccount": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "BUSINESS",
              "PERSONAL"
            ]
          },
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1
          },
          "country": {
            "type": "string",
            "minLength": 1
          },
          "businessName": {
            "type": "string",
            "minLength": 1
          },
          "businessRegistrationNumber": {
            "type": "string",
            "minLength": 1
          },
          "taxId": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "postalCode": {
            "type": "string",
            "minLength": 1
          },
          "city": {
            "type": "string",
            "minLength": 1
          },
          "tradeRole": {
            "type": "string",
            "enum": [
              "IMPORTER",
              "EXPORTER",
              "FREIGHT_FORWARDER",
              "CUSTOMS_BROKER",
              "MANUFACTURER",
              "DISTRIBUTOR",
              "TRADER"
            ]
          }
        },
        "id": "UpdateAccount"
      }
    }
  },
  "security": [
    {
      "clerkSession": []
    }
  ]
}